requirements.txt

Flask-Caching==2.1.0
# celery==5.3.4
# redis==5.0.1
Flask
Flask-SQLAlchemy
Flask-Security-Too
Jinja2
# Flask-Bcrypt
celery[redis]
Werkzeug==3.1.3
flask_restful
flask-cors
setuptools
